Get Data representation for `Writable` (#787)
* get string representation of pbxproj

* made the outputSettings with `default` value

* reverted change to parameter

* Refactored code

* Added dataRepresentation() to Writable

* fixed issue with code

+++ Sources/XcodeProj/Objects/Project/PBXProj.swift
@@ -284,13 +284,24 @@ extension PBXProj: Equatable {
 // MARK: - Writable
 
 extension PBXProj: Writable {
+    public func dataRepresentation(outputSettings: PBXOutputSettings) throws -> Data? {
+        let encoder = PBXProjEncoder(outputSettings: outputSettings)
+        return try encoder.encode(proj: self).data(using: .utf8)
+    }
+
+    public func dataRepresentation() throws -> Data? {
+        let encoder = PBXProjEncoder(outputSettings: PBXOutputSettings())
+        return try encoder.encode(proj: self).data(using: .utf8)
+    }
+    
     public func write(path: Path, override: Bool) throws {
         try write(path: path, override: override, outputSettings: PBXOutputSettings())
     }
 
     public func write(path: Path, override: Bool, outputSettings: PBXOutputSettings) throws {
-        let encoder = PBXProjEncoder(outputSettings: outputSettings)
-        let output = try encoder.encode(proj: self)
+        guard let output = try dataRepresentation(outputSettings: outputSettings) else {
+            return
+        }
         if override, path.exists {
             try path.delete()
         }
